requirements.txt
amqp==2.6.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.7
Django==4.0.4
flower==0.9.7
humanize==4.1.0
kombu==4.6.11
prometheus-client==0.8.0
pytz==2022.1
redis==3.5.3
sqlparse==0.4.2
tornado==6.1
vine==1.3.0
djangorestframework-3.13.1